LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/canvas/offscreen/manual/the-offscreen-canvas/offscreencanvas.transfer.to.imagebitmap.w-expected.txt

CONSOLE MESSAGE: InvalidStateError: The object is in an invalid state.

Harness Error (FAIL), message = InvalidStateError: The object is in an invalid state.

FAIL Test that transferToImageBitmap returns an ImageBitmap with correct width and height in a worker assert_unreached: error Reached unreachable code
PASS Test that transferToImageBitmap returns an ImageBitmap with correct color in a worker
PASS Test that call transferToImageBitmap twice returns an ImageBitmap with correct color in a worker
FAIL Test that call transferToImageBitmap twice on a alpha-disabled context returns an ImageBitmap with correct color in a worker assert_array_equals: expected property 3 to be 255 but got 0 (expected array [0, 0, 0, 255] got object "0,0,0,0")
PASS Test that transferToImageBitmap won't change context's property in a worker
PASS Test that call transferToImageBitmap preserves transform in a worker
PASS Test that call transferToImageBitmap on a detached OffscreenCanvas throws an exception in a worker
PASS Test that call transferToImageBitmap without a context throws an exception in a worker
